pyproject.toml
[tool.poetry]
name = "sc_blog"
version = "0.1.0"
description = ""
authors = ["Neil Millard <neil@neilmillard.com>"]

[tool.poetry.dependencies]
python = "^3.8"
django = "^3.0.8"
python-decouple = "^3.3"
dj-database-url = "^0.5.0"
psycopg2-binary = "^2.8.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
